# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EngChat API",
        "description": "API documentation for EngChat - Customer Service Platform",
        "contact": {
            "name": "EngChat Development Team",
            "email": "dev@engchat.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "EngChat API Server"
        }
    ],
    "paths": {
        "/api/conversations": {
            "get": {
                "tags": [
                    "Conversations"
                ],
                "summary": "List conversations",
                "description": "Get a list of conversations for the authenticated user",
                "operationId": "getConversations",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by conversation status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "open",
                                "pending",
                                "closed"
                            ]
                        }
                    },
                    {
                        "name": "priority",
                        "in": "query",
                        "description": "Filter by priority",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "low",
                                "medium",
                                "high",
                                "urgent"
                            ]
                        }
                    },
                    {
                        "name": "channel_id",
                        "in": "query",
                        "description": "Filter by channel ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Conversation"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Conversations"
                ],
                "summary": "Create a new conversation",
                "description": "Creates a new conversation",
                "operationId": "createConversation",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "contact_id",
                                    "channel_id",
                                    "subject"
                                ],
                                "properties": {
                                    "contact_id": {
                                        "description": "Contact ID",
                                        "type": "integer"
                                    },
                                    "channel_id": {
                                        "description": "Channel ID",
                                        "type": "integer"
                                    },
                                    "subject": {
                                        "description": "Conversation subject",
                                        "type": "string"
                                    },
                                    "priority": {
                                        "description": "Priority level",
                                        "type": "string",
                                        "enum": [
                                            "low",
                                            "medium",
                                            "high",
                                            "urgent"
                                        ]
                                    },
                                    "category_id": {
                                        "description": "Category ID",
                                        "type": "integer"
                                    },
                                    "description": {
                                        "description": "Initial message/description",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Conversation created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Conversation"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Channel": {
                "title": "Channel",
                "description": "Model representing a communication channel (WhatsApp, Telegram, etc.)",
                "required": [
                    "name",
                    "type"
                ],
                "properties": {
                    "id": {
                        "description": "Channel ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "description": "Channel name",
                        "type": "string",
                        "example": "WhatsApp Business"
                    },
                    "type": {
                        "description": "Channel type",
                        "type": "string",
                        "enum": [
                            "whatsapp",
                            "telegram",
                            "web",
                            "facebook",
                            "instagram"
                        ],
                        "example": "whatsapp"
                    },
                    "configuration": {
                        "description": "Channel configuration",
                        "type": "object",
                        "example": {
                            "phone": "554133808848",
                            "api_version": "v17.0"
                        },
                        "nullable": true
                    },
                    "is_active": {
                        "description": "Whether channel is active",
                        "type": "boolean",
                        "example": true
                    },
                    "priority": {
                        "description": "Channel priority",
                        "type": "integer",
                        "example": 1
                    },
                    "working_hours": {
                        "description": "Working hours configuration",
                        "type": "object",
                        "nullable": true
                    },
                    "auto_response_enabled": {
                        "description": "Whether auto-response is enabled",
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Contact": {
                "title": "Contact",
                "description": "Model representing a customer contact",
                "required": [
                    "name"
                ],
                "properties": {
                    "id": {
                        "description": "Contact ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "description": "Contact full name",
                        "type": "string",
                        "example": "Maria Silva"
                    },
                    "email": {
                        "description": "Contact email",
                        "type": "string",
                        "format": "email",
                        "example": "maria@example.com",
                        "nullable": true
                    },
                    "phone": {
                        "description": "Contact phone number",
                        "type": "string",
                        "example": "+5511999999999",
                        "nullable": true
                    },
                    "display_name": {
                        "description": "Display name for chat",
                        "type": "string",
                        "example": "Maria",
                        "nullable": true
                    },
                    "company": {
                        "description": "Company name",
                        "type": "string",
                        "example": "Empresa XYZ",
                        "nullable": true
                    },
                    "document": {
                        "description": "Document number (CPF, CNPJ)",
                        "type": "string",
                        "example": "123.456.789-10",
                        "nullable": true
                    },
                    "tags": {
                        "description": "Contact tags",
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "vip",
                            "premium"
                        ],
                        "nullable": true
                    },
                    "priority": {
                        "description": "Contact priority",
                        "type": "string",
                        "enum": [
                            "low",
                            "medium",
                            "high",
                            "urgent"
                        ],
                        "example": "medium"
                    },
                    "blacklisted": {
                        "description": "Whether contact is blacklisted",
                        "type": "boolean",
                        "example": false
                    },
                    "blacklist_reason": {
                        "description": "Reason for blacklisting",
                        "type": "string",
                        "nullable": true
                    },
                    "preferred_language": {
                        "description": "Preferred language",
                        "type": "string",
                        "example": "pt-BR"
                    },
                    "timezone": {
                        "description": "Contact timezone",
                        "type": "string",
                        "example": "America/Sao_Paulo"
                    },
                    "last_interaction": {
                        "description": "Last interaction timestamp",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "total_interactions": {
                        "description": "Total number of interactions",
                        "type": "integer",
                        "example": 15
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Conversation": {
                "title": "Conversation",
                "description": "Model representing a customer conversation",
                "required": [
                    "id",
                    "contact_id",
                    "channel_id",
                    "status",
                    "subject"
                ],
                "properties": {
                    "id": {
                        "description": "Conversation ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "contact_id": {
                        "description": "Contact ID",
                        "type": "integer",
                        "example": 1
                    },
                    "channel_id": {
                        "description": "Channel ID",
                        "type": "integer",
                        "example": 1
                    },
                    "category_id": {
                        "description": "Category ID",
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "assigned_to": {
                        "description": "Assigned user ID",
                        "type": "integer",
                        "example": 5,
                        "nullable": true
                    },
                    "status": {
                        "description": "Conversation status",
                        "type": "string",
                        "enum": [
                            "open",
                            "pending",
                            "closed"
                        ],
                        "example": "open"
                    },
                    "priority": {
                        "description": "Priority level",
                        "type": "string",
                        "enum": [
                            "low",
                            "medium",
                            "high",
                            "urgent"
                        ],
                        "example": "medium"
                    },
                    "subject": {
                        "description": "Conversation subject",
                        "type": "string",
                        "example": "Problema com produto"
                    },
                    "last_message_at": {
                        "description": "Last message timestamp",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "closed_at": {
                        "description": "Closure timestamp",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "closed_by": {
                        "description": "User who closed the conversation",
                        "type": "integer",
                        "nullable": true
                    },
                    "metadata": {
                        "description": "Additional metadata",
                        "type": "object",
                        "nullable": true
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "description": "Deletion timestamp",
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "contact": {
                        "$ref": "#/components/schemas/Contact"
                    },
                    "channel": {
                        "$ref": "#/components/schemas/Channel"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use Bearer token for authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}